refactor(editor): isolate project file lifecycle
Move formal-file state, remembered save targets, dirty replacement guards, previous-project restoration, recovery-session actions, Project opening, and shelf opening into useProjectFileLifecycle. App now supplies one bounded installation callback for resetting canvas, Agent, import, and navigation state after a validated replacement.

This removes 448 net lines from App and keeps recovery plus formal-file transitions in one ownership boundary.

Validation: pnpm ci:static; App, project-file-service, and recovery-coordinator units (50); complete project-file E2E (15); focused recovery/manual E2E (5); git diff --check.

No test files changed because the existing project-file and recovery contracts directly cover every moved transition.

Unrelated user-owned probe-conflicts.mjs remained untracked and untouched.

Test-Impact: no-test-change — behavior-preserving ownership move covered by 50 focused units and 20 existing browser scenarios.
